@@ -1,6 +1,5 @@
-module.exports = {
-  plugins: [
-    require('tailwindcss'),
-    require('autoprefixer'),
-  ]
+export default {
+  plugins: {
+    '@tailwindcss/postcss': {},
+  },
 }
