@@ -1,7 +1,13 @@
-import React from 'react';
-import Checkbox from '../../shared/Checkbox';
+import { RecommendationTypeProps } from '../../../types'
+import Checkbox from '../../shared/Checkbox'
 
-function RecommendationType({ onRecommendationTypeChange }) {
+interface Props extends RecommendationTypeProps {
+  onRecommendationTypeChange: (
+    type: 'SingleProduct' | 'MultipleProducts',
+  ) => void
+}
+
+function RecommendationType({ onRecommendationTypeChange }: Props) {
   return (
     <div className="mb-4">
       <h2 className="text-lg font-bold mb-2">Tipo de Recomendação:</h2>
@@ -13,7 +19,9 @@ function RecommendationType({ onRecommendationTypeChange }) {
           onChange={() => onRecommendationTypeChange('SingleProduct')}
           className="mr-2"
         />
-        <label htmlFor="SingleProduct" className="mr-4">Produto Único</label>
+        <label htmlFor="SingleProduct" className="mr-4">
+          Produto Único
+        </label>
         <Checkbox
           type="radio"
           name="recommendationType"
@@ -24,7 +32,7 @@ function RecommendationType({ onRecommendationTypeChange }) {
         <label htmlFor="MultipleProducts">Múltiplos Produtos</label>
       </div>
     </div>
-  );
+  )
 }
 
-export default RecommendationType;
+export default RecommendationType
