@@ -1,25 +1,26 @@
-import recommendationService from './recommendation.service';
-import mockProducts from '../mocks/mockProducts';
+import mockProducts from '../mocks/mockProducts'
+import recommendationService from './recommendation.service'
+import { FormData } from '../types'
 
 describe('recommendationService', () => {
   test('Retorna recomendação correta para SingleProduct com base nas preferências selecionadas', () => {
-    const formData = {
+    const formData: FormData = {
       selectedPreferences: ['Integração com chatbots'],
       selectedFeatures: ['Chat ao vivo e mensagens automatizadas'],
       selectedRecommendationType: 'SingleProduct',
-    };
+    }
 
     const recommendations = recommendationService.getRecommendations(
       formData,
-      mockProducts
-    );
+      mockProducts,
+    )
 
-    expect(recommendations).toHaveLength(1);
-    expect(recommendations[0].name).toBe('RD Conversas');
-  });
+    expect(recommendations).toHaveLength(1)
+    expect(recommendations[0]?.name).toBe('RD Conversas')
+  })
 
   test('Retorna recomendações corretas para MultipleProducts com base nas preferências selecionadas', () => {
-    const formData = {
+    const formData: FormData = {
       selectedPreferences: [
         'Integração fácil com ferramentas de e-mail',
         'Personalização de funis de vendas',
@@ -30,22 +31,22 @@ describe('recommendationService', () => {
         'Rastreamento de comportamento do usuário',
       ],
       selectedRecommendationType: 'MultipleProducts',
-    };
+    }
 
     const recommendations = recommendationService.getRecommendations(
       formData,
-      mockProducts
-    );
+      mockProducts,
+    )
 
-    expect(recommendations).toHaveLength(2);
+    expect(recommendations).toHaveLength(2)
     expect(recommendations.map((product) => product.name)).toEqual([
       'RD Station CRM',
       'RD Station Marketing',
-    ]);
-  });
+    ])
+  })
 
   test('Retorna apenas um produto para SingleProduct com mais de um produto de match', () => {
-    const formData = {
+    const formData: FormData = {
       selectedPreferences: [
         'Integração fácil com ferramentas de e-mail',
         'Automação de marketing',
@@ -55,29 +56,33 @@ describe('recommendationService', () => {
         'Rastreamento de comportamento do usuário',
       ],
       selectedRecommendationType: 'SingleProduct',
-    };
+    }
 
     const recommendations = recommendationService.getRecommendations(
       formData,
-      mockProducts
-    );
+      mockProducts,
+    )
 
-    expect(recommendations).toHaveLength(1);
-    expect(recommendations[0].name).toBe('RD Station Marketing');
-  });
+    expect(recommendations).toHaveLength(1)
+    expect(recommendations[0]?.name).toBe('RD Station Marketing')
+  })
 
   test('Retorna o último match em caso de empate para SingleProduct', () => {
-    const formData = {
-      selectedPreferences: ['Automação de marketing', 'Integração com chatbots'],
+    const formData: FormData = {
+      selectedPreferences: [
+        'Automação de marketing',
+        'Integração com chatbots',
+      ],
+      selectedFeatures: [],
       selectedRecommendationType: 'SingleProduct',
-    };
+    }
 
     const recommendations = recommendationService.getRecommendations(
       formData,
-      mockProducts
-    );
+      mockProducts,
+    )
 
-    expect(recommendations).toHaveLength(1);
-    expect(recommendations[0].name).toBe('RD Conversas');
-  });
-});
+    expect(recommendations).toHaveLength(1)
+    expect(recommendations[0]?.name).toBe('RD Conversas')
+  })
+})
